requirements.txt
# Core pipeline dependencies
pypdfium2
python-pptx
scikit-learn
